requirements.txt
faiss-cpu
numpy
pypdf
sentence-transformers
transformers
torch
